:white_check_mark: Creates tests to model mapping.
